Revert "Feat/annotation model"
